use new smoothing method and update the income mountains
the previous method of smoothing was ad hoc and doesn't retain the
movements of the shapes, which means that the poverty rates
calculated from the shapes are not following the original data.

I so try to use a better way to handle the smoothing.

Now the smoothing was done by applying a Savitzky-Golay filter with
a adaptive window size, which is determined by the Full width of half maximum
for a shape. See etl/notebooks/smoothing_method_3.py for more details.